Unify logsdb index settings providers (#118342)
* Unify logsdb index settings providers

* restore diff

* rename method

(cherry picked from commit c7925957f91fc4ba5a2b834f0e368dd4864f8cf4)

# Conflicts:
#	x-pack/plugin/logsdb/src/main/java/org/elasticsearch/xpack/logsdb/LogsDBPlugin.java
